Move store/search to core/lib/search
