+++ core/lib/search/adapter/api/elasticsearchQuery.js
