+++ core/lib/search/adapter/graphql/gqlQuery.js
