+++ core/lib/search/adapter/graphql/queries/categories.gql
