+++ core/lib/search/adapter/graphql/queries/cmsBlock.gql
