+++ core/lib/search/adapter/graphql/queries/cmsHierarchy.gql
