+++ core/lib/search/adapter/graphql/queries/cmsPage.gql
