+++ core/lib/search/adapter/graphql/queries/customAttributeMetadata.gql
