+++ core/lib/search/adapter/graphql/queries/products.gql
