+++ core/lib/search/adapter/graphql/queries/reviews.gql
