+++ core/lib/search/adapter/graphql/queries/taxrule.gql
