+++ core/lib/search/adapter/searchAdapterFactory.js
