+++ core/lib/search/boost.js
