+++ core/lib/search/mapping.js
