+++ core/lib/search/searchQuery.js
