+++ core/modules/catalog/store/category/actions.ts
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import { ActionTree } from 'vuex'
 import * as types from './mutation-types'
-import { quickSearchByQuery } from '@vue-storefront/store/lib/search'
+import { quickSearchByQuery } from '@vue-storefront/core/lib/search'
 import { entityKeyName } from '@vue-storefront/store/lib/entities'
 import rootStore from '@vue-storefront/store'
 import i18n from '@vue-storefront/i18n'
@@ -11,7 +11,7 @@ import toString from 'lodash-es/toString'
 import { optionLabel } from '../../helpers/optionLabel'
 import RootState from '@vue-storefront/store/types/RootState'
 import CategoryState from '../../types/CategoryState'
-import SearchQuery from '@vue-storefront/store/lib/search/searchQuery'
+import SearchQuery from '@vue-storefront/core/lib/search/searchQuery'
 import { currentStoreView } from '@vue-storefront/store/lib/multistore'
 
 const actions: ActionTree<CategoryState, RootState> = {
