+++ core/modules/catalog/store/product/actions.ts
@@ -13,10 +13,10 @@ import { configureProductAsync,
   getMediaGallery, 
   configurableChildrenImages, 
   attributeImages } from '../../helpers'
-import SearchQuery from '@vue-storefront/store/lib/search/searchQuery'
+import SearchQuery from '@vue-storefront/core/lib/search/searchQuery'
 import { entityKeyName } from '@vue-storefront/store/lib/entities'
 import { optionLabel } from '../../helpers/optionLabel'
-import { quickSearchByQuery, isOnline } from '@vue-storefront/store/lib/search'
+import { quickSearchByQuery, isOnline } from '@vue-storefront/core/lib/search'
 import omit from 'lodash-es/omit'
 import trim from 'lodash-es/trim'
 import uniqBy from  'lodash-es/uniqBy'
