+++ core/modules/cms/store/hierarchy/actions.ts
@@ -1,6 +1,6 @@
 import { ActionTree } from "vuex"
-import { quickSearchByQuery } from '@vue-storefront/store/lib/search'
-import SearchQuery from '@vue-storefront/store/lib/search/searchQuery'
+import { quickSearchByQuery } from '@vue-storefront/core/lib/search'
+import SearchQuery from '@vue-storefront/core/lib/search/searchQuery'
 import RootState from '@vue-storefront/store/types/RootState';
 import CmsHierarchyState from "../../types/CmsHierarchyState"
